+++ R/drugSensitivity.R
@@ -445,11 +445,11 @@ writeExpressionDrugSensitivityCorHDF5 <- function(
 #' @examples
 #' listExpressionDrugSensitivityAssociation()
 listExpressionDrugSensitivityAssociation <- function(url=FALSE) {
-    options <- c(
-        "GDSC 7"="5q0dazbtnpojw2m/expressionDrugSensitivityCorGDSC7.rds",
-        "CTRP 2.1"="zj53pxwiwdwo133/expressionDrugSensitivityCorCTRP2.1.rds",
-        "NCI60"="20ko9lyyyoilfz6/expressionDrugSensitivityCorNCI60.h5")
-    link <- sprintf("https://www.dropbox.com/s/%s?raw=1", options)
+    options <- c("GDSC 7"="expressionDrugSensitivityCorGDSC7.qs",
+                 "CTRP 2.1"="expressionDrugSensitivityCorCTRP2.1.qs",
+                 "NCI60"="expressionDrugSensitivityCorNCI60.h5")
+    link <- file.path("https://compbio.imm.medicina.ulisboa.pt/public/cTRAP",
+                      options)
     names(link) <- names(options)
 
     res <- link
@@ -601,6 +601,7 @@ dim.expressionDrugSensitivityAssociation <- function(x) {
 #' @export
 #'
 #' @importFrom tools file_ext
+#' @importFrom qs qread
 #'
 #' @examples
 #' gdsc <- listExpressionDrugSensitivityAssociation()[[1]]
@@ -616,17 +617,22 @@ loadExpressionDrugSensitivityAssociation <- function(
     if (!is.null(path)) file <- file.path(path, file)
     downloadIfNotFound(link, file)
     message(sprintf("Loading data from %s...", file))
-    if (file_ext(file) == "rds") {
+    if (file_ext(file) == "h5") {
+        cor <- readExpressionDrugSensitivityCorHDF5(file, rows=rows, cols=cols,
+                                                    loadValues=loadValues)
+    } else {
         if (is.null(cols)) cols <- TRUE
         if (is.null(rows)) rows <- TRUE
-        res <- readRDS(file)
+        
+        if (file_ext(file) == "rds") {
+            res <- readRDS(file)
+        } else if (file_ext(file) == "qs") {
+            res <- qread(file)
+        } 
         cor <- res[rows, cols, drop=FALSE]
         attrs <- attributes(res)
         attrs <- attrs[!names(attrs) %in% names(attributes(cor))]
         attributes(cor) <- c(attributes(cor), attrs)
-    } else {
-        cor <- readExpressionDrugSensitivityCorHDF5(file, rows=rows, cols=cols,
-                                                    loadValues=loadValues)
     }
     attr(cor, "filename") <- normalizePath(file)
     class(cor) <- c("expressionDrugSensitivityAssociation", class(cor))